+++ Sanctuary.Census/Abstractions/Database/ICollectionsContext.cs
@@ -213,6 +213,14 @@ public interface ICollectionsContext
     Task UpsertVehicleLoadoutSlotsAsync(IEnumerable<VehicleLoadoutSlot> collection, CancellationToken ct = default);
 
     /// <summary>
+    /// Upserts the <see cref="VehicleSkillSet"/> collection.
+    /// </summary>
+    /// <param name="collection">The collection.</param>
+    /// <param name="ct">A <see cref="CancellationToken"/> that can be used to stop the operation.</param>
+    /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
+    Task UpsertVehicleSkillSetsAsync(IEnumerable<VehicleSkillSet> collection, CancellationToken ct = default);
+
+    /// <summary>
     /// Upserts the <see cref="Weapon"/> collection.
     /// </summary>
     /// <param name="collection">The collection.</param>